尝试3

+++ parser.cpp
@@ -16,7 +16,7 @@ QString Parser::parserfirst(QString text)
     QRegularExpression regex_showdatabases("\\bshow\\s+databases;", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_selectdatabase("\\bselect\\s+database\\(\\);", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_use("\\buse\\s+\\w+;", QRegularExpression::CaseInsensitiveOption);
-    QRegularExpression regex_drop("\\bdrop\\s+\\w+;", QRegularExpression::CaseInsensitiveOption);
+    QRegularExpression regex_drop("\\bdrop\\s+\\w+\\s+\\w+;", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_delete("\\bdelete\\s+\\w+;", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_insertinto("\\bINSERT\\s+INTO\\s+\\w+\\s*\\([^\\)]+\\)\\s*VALUES\\s*\\([^\\)]+\\);", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_createtable("\\bCREATE\\s+TABLE\\s+(\\w+)\\s*\\(([^;]+)\\);", QRegularExpression::CaseInsensitiveOption);
@@ -24,7 +24,6 @@ QString Parser::parserfirst(QString text)
     QRegularExpression regex_alter("\\balter\\b", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_show("\\bshow\\b", QRegularExpression::CaseInsensitiveOption);
     QRegularExpression regex_grant("\\bgrant\\b", QRegularExpression::CaseInsensitiveOption);
-
     QRegularExpression regex_select("\\bselect\\s+\\S+\\s+from\\s+\\w+.*;", QRegularExpression::CaseInsensitiveOption);
 
 
@@ -98,15 +97,27 @@ QString Parser::parserfirst(QString text)
     {
         // Retrieve the matched QString
         QString matchedString = match_drop.captured();
-        //获取create database dbname三个分开的单词
-        QStringList words = matchedString.split(QRegExp("\\s+"), QString::SkipEmptyParts);
-        QString tablename = words.at(1);
-        //去掉分号
-        tablename.chop(1);
-        //调用相应函数
-        string tname = tablename.toStdString();
-        Table a(tname,db);
-        a.dropTable();
+        //获取drop database/table name三个分开的单词
+        QStringList words = matchedString.split(QRegExp("\\s+"), QString::SkipEmptyParts);        
+        QString type = words.at(1);
+        if(type == "database")
+        {
+             QString dbname = words.at(2);
+             //去掉分号
+             dbname.chop(1);
+             //调用dropdb函数
+             db->dropdb(dbname);
+        }
+        if(type == "table")
+        {
+             QString tablename = words.at(2);
+             //去掉分号
+             tablename.chop(1);
+             //调用相应函数dropTable
+             string tname = tablename.toStdString();
+             Table a(tname,db);
+             a.dropTable();
+        }
 
 
 
@@ -135,20 +146,18 @@ QString Parser::parserfirst(QString text)
 
 
 
-    //insertinto
+    //insert into t1 (sno,sname) values(12,22);
     QRegularExpressionMatch match_insertinto = regex_insertinto.match(text);
 
-    if (match_insertinto.hasMatch()) {
-
+    if (match_insertinto.hasMatch())
+    {
         QString matchedString = match_insertinto.captured();
-        QStringList words = matchedString.split(QRegExp("\\s+"), QString::SkipEmptyParts);
-        //get table name
+        QStringList words = matchedString.split(QRegExp("\\s+"), QString::SkipEmptyParts);        
         QString tablename = words[2];
+        //剔除括号
         QRegularExpression regex_presenthesis("\\([^\\)]+\\)");
         QRegularExpressionMatchIterator matchIterator = regex_presenthesis.globalMatch(matchedString);
-
         QString attribute[2];
-
         if (!matchIterator.hasNext())
         {
             qDebug() << "error";
@@ -159,8 +168,8 @@ QString Parser::parserfirst(QString text)
             QString matchedText = match.captured(0);
             matchedText.remove(0, 1);
             matchedText.remove(matchedText.length() - 1, 1);
-            //1, 'John Doe', 'IT', 50000 get every word without ,
 
+            //
             QRegularExpression regex("\\b\\w+\\b");
             QRegularExpressionMatchIterator matchIterator = regex.globalMatch(matchedText);
             while (matchIterator.hasNext()) {
@@ -222,11 +231,11 @@ QString Parser::parserfirst(QString text)
 
     }
 
+    // select aas,sds,asf from table /where a = b;
     QRegularExpressionMatch match_select = regex_select.match(text);
-
     if (match_select.hasMatch())
     {
-        //select aas,sds,asf from table where a = b;
+
         QString matchedString = match_select.captured();
         //获取分开的单词
         QStringList words = matchedString.split(QRegExp("\\s+"), QString::SkipEmptyParts);
@@ -242,12 +251,29 @@ QString Parser::parserfirst(QString text)
         {
             //逗号分割
             QStringList columnList = secondpos.split(",");
-            //重复工作
-            QString column_array[columnList.size()];
-            for (int i = 0; i < columnList.size(); i++) {
-                column_array[i] = columnList[i];
+            //重复工作qstring改string
+            int size = columnList.size();
+            vector <string> column_array(size);
+            for (int i = 0; i < size; i++) {
+                column_array[i] = columnList[i].toStdString();
+            }
+            //看第四个位置;或第五个位置为where，或者嵌套查询
+            if(fourthpos.back() == ';')
+            {
+                fourthpos.chop(1);
+                string tablename =fourthpos.toStdString();
+                Table tb(tablename,db);
+                tb.selectfrom(column_array,tablename);
+            }
+            if(fifthpos == "where")
+            {
+
             }
         }
+
+
+
+
